refactor code for testing
