+++ condow_core/CHANGELOG.md
@@ -4,18 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
-
-### CHANGED
-
-- **BREAKING** removed `time` parameter from `Probe::chunk_completed` (was inaccurate and had a negative performance impact)
-- improved performance on sequential downloads
-
-## 0.17.1 - 2022-05-06
+## 0.17.1 - unreleased
 
 ### CHANGED
 
 - `Probe` is handled with static dispatch internally
+- Deprecated `Probe::chunk_completed` (time parameter was inaccurate and had a negative performance impact). Use new method `chunk_received` instead.
+- improved performance on sequential downloads
 
 ## 0.17.0 - 2022-05-05
